Add a Write trait
This will be useful for TLS later on.
